Move infinite types to swr/infinite
