+++ src/index.ts
@@ -8,8 +8,6 @@ export default useSWR
 // Types
 export {
   SWRConfiguration,
-  SWRInfiniteConfiguration,
-  SWRInfiniteResponse,
   Revalidator,
   RevalidatorOptions,
   Key,
@@ -22,8 +20,6 @@ export {
   Middleware,
   // Legacy, for backwards compatibility
   ConfigInterface,
-  SWRInfiniteConfigInterface,
-  SWRInfiniteResponseInterface,
   revalidateType,
   RevalidateOptionInterface,
   keyInterface,
